Start - add music & more krill. Vignette speed
Changed the speed of dying and how long it takes to die after the vignette is full. Added more krill at the start of the level. Changed the struggle sound slightly to only play natural sound, but on button press. Will have to turn up the sound a bit and add other sounds. Added music to the death screen and set up for the title screen.