A lot of sound stuff
Added net struggling sounds and a more realistic sound when full from eating. Added making the audio quieter in later zones, and added fade-ins and outs to the movement audio. Added another music file to the tutorial screen. Got rid of unused sounds. Changed the text for the regular death screen. Made small tweaks to a lot of audios that may be further adjusted.